+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### **2.0.3 (unreleased)** 
 
-- 🛠️ Fixed resetting form repeating some fieldset items due to a bug in the removeAllButton function.
+- 🛠️ Fixed resetting form repeating some fieldset items due to a bug in the removeAllButtons function.
 - ⚙️ Removed unnecessary visibility toggling when opening/closing popups.
 - ⚙️ Moved popup CSS classes to its own CSS module instead of the common module.
 - ⚙️ Removed redundant isOpen flag from popup component.